+++ integration/ciaolgt.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ## 
 ##   Integration script for Ciao
-##   Last updated on September 30, 2022
+##   Last updated on March 19, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -24,76 +24,7 @@
 #############################################################################
 
 
-if ! [ "$LOGTALKHOME" ]; then
-	echo "The environment variable LOGTALKHOME should be defined first"
-	echo "pointing to your Logtalk installation directory!"
-	echo "Trying the default locations for the Logtalk installation..."
-	if [ -d "/usr/local/share/logtalk" ]; then
-		LOGTALKHOME=/usr/local/share/logtalk
-		echo "... using Logtalk installation found at /usr/local/share/logtalk"
-	elif [ -d "/usr/share/logtalk" ]; then
-		LOGTALKHOME=/usr/share/logtalk
-		echo "... using Logtalk installation found at /usr/share/logtalk"
-	elif [ -d "/opt/local/share/logtalk" ]; then
-		LOGTALKHOME=/opt/local/share/logtalk
-		echo "... using Logtalk installation found at /opt/local/share/logtalk"
-	elif [ -d "/opt/share/logtalk" ]; then
-		LOGTALKHOME=/opt/share/logtalk
-		echo "... using Logtalk installation found at /opt/share/logtalk"
-	elif [ -d "/opt/homebrew/share/logtalk" ]; then
-		LOGTALKHOME=/opt/homebrew/share/logtalk
-		echo "... using Logtalk installation found at /opt/homebrew/share/logtalk"
-	elif [ -d "$HOME/share/logtalk" ]; then
-		LOGTALKHOME="$HOME/share/logtalk"
-		echo "... using Logtalk installation found at $HOME/share/logtalk"
-	elif [ -f "$( cd "$( dirname "$0" )" && pwd )/../core/core.pl" ]; then
-		LOGTALKHOME="$( cd "$( dirname "$0" )" && pwd )/.."
-		echo "... using Logtalk installation found at $( cd "$( dirname "$0" )" && pwd )/.."
-	else
-		echo "... unable to locate Logtalk installation directory!" >&2
-		echo
-		exit 1
-	fi
-	echo
-	export LOGTALKHOME=$LOGTALKHOME
-elif ! [ -d "$LOGTALKHOME" ]; then
-	echo "The environment variable LOGTALKHOME points to a non-existing directory!" >&2
-	echo "Its current value is: $LOGTALKHOME" >&2
-	echo "The variable must be set to your Logtalk installation directory!" >&2
-	echo
-	exit 1
-fi
-
-if ! [ "$LOGTALKUSER" ]; then
-	echo "The environment variable LOGTALKUSER should be defined first, pointing"
-	echo "to your Logtalk user directory!"
-	echo "Trying the default location for the Logtalk user directory..."
-	echo
-	export LOGTALKUSER=$HOME/logtalk
-fi
-
-if [ -d "$LOGTALKUSER" ]; then
-	if ! [ -f "$LOGTALKUSER/VERSION.txt" ]; then
-		echo "Cannot find version information in the Logtalk user directory at $LOGTALKUSER!"
-		echo "Creating an up-to-date Logtalk user directory..."
-		logtalk_user_setup
-	else
-		system_version=$(cat "$LOGTALKHOME/VERSION.txt")
-		user_version=$(cat "$LOGTALKUSER/VERSION.txt")
-		if [ "$user_version" \< "$system_version" ]; then
-			echo "Logtalk user directory at $LOGTALKUSER is outdated: "
-			echo "    $user_version < $system_version"
-			echo "Creating an up-to-date Logtalk user directory..."
-			logtalk_user_setup
-		fi
-	fi
-else
-	echo "Cannot find \$LOGTALKUSER directory! Creating a new Logtalk user directory"
-	echo "by running the \"logtalk_user_setup\" shell script:"
-	logtalk_user_setup
-fi
-
-LOGTALK_STARTUP_DIRECTORY=$(pwd)
-export LOGTALK_STARTUP_DIRECTORY
+source "$(dirname "$0")/logtalk_setup_env.sh"
+setup_logtalk_env || exit 1
 
 exec ciaosh --iso-strict -l "$LOGTALKHOME/integration/logtalk_ciao.pl" "$@"